initial splitting out of study form components
