+++ web/src/types/studies.d.ts
@@ -1,0 +1,29 @@
+type StudyFormData = {
+  title: string;
+  description: string | undefined;
+  owner: string;
+  additionalStudyAdminUsernames: { value: string }[];
+  dataControllerOrganisation: string;
+  cagReference: string | undefined;
+  dataProtectionPrefix: string;
+  dataProtectionDate: string;
+  dataProtectionId: string;
+  dataProtectionNumber: string | undefined | null;
+  nhsEnglandReference: string | undefined;
+  irasId: string | undefined;
+  involvesUclSponsorship: boolean | null;
+  involvesCag: boolean | null;
+  involvesEthicsApproval: boolean | null;
+  involvesHraApproval: boolean | null;
+  requiresDbs: boolean | null;
+  isDataProtectionOfficeRegistered: boolean | null;
+  involvesThirdParty: boolean | null;
+  involvesExternalUsers: boolean | null;
+  involvesParticipantConsent: boolean | null;
+  involvesIndirectDataCollection: boolean | null;
+  involvesDataProcessingOutsideEea: boolean | null;
+  isNhsAssociated: boolean | null;
+  involvesNhsEngland: boolean | null;
+  involvesMnca: boolean | null;
+  requiresDspt: boolean | null;
+};
